@@ -1,34 +1,29 @@
 name: flutter_wordpress_app
-description: Flutter news app for wordpress platform
-
+description: "Flutter news app for wordpress platform"
 version: 2.0.0+5
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ^3.6.0
 
 dependencies:
   flutter:
     sdk: flutter
-
-  cupertino_icons: ^1.0.0
-  url_launcher: ^6.0.11
-  http: ^0.13.3
-  html: ^0.15.0
-  flutter_html: ^2.1.2
-  dio: ^4.0.0
-  loading: ^1.0.2
-  intl: ^0.17.0
-  share: ^2.0.4
-  path_provider: ^2.0.5
-  sqflite: ^2.0.0+4
-  shared_preferences: ^2.0.8
-  flutter_widget_from_html: ^0.7.0
+  url_launcher: ^6.3.1
+  http: ^1.2.2
+  html: ^0.15.4
+  flutter_html: ^2.2.1
+  dio: ^5.7.0
+  intl: ^0.20.1
+  share_plus: ^6.3.4
+  path_provider: ^2.1.5
+  sqflite: ^2.4.1
+  shared_preferences: ^2.3.4
+  flutter_widget_from_html: ^0.9.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-
-  flutter_launcher_icons: ^0.9.2
+  flutter_lints: ^5.0.0
 
 flutter_icons:
   android: "launcher_icon"
@@ -36,10 +31,6 @@ flutter_icons:
   image_path: assets/launcher_icon.png
 
 flutter:
-  # The following line ensures that the Material Icons font is
-  # included with your application, so that you can use the icons in
-  # the material Icons class.
-
   uses-material-design: true
 
   assets:
