@@ -11,9 +11,9 @@ import 'package:http/http.dart' as http;
 class CategoryArticles extends StatefulWidget {
   final int id;
   final String name;
-  CategoryArticles(this.id, this.name, {Key? key}) : super(key: key);
+  const CategoryArticles(this.id, this.name, {super.key});
   @override
-  _CategoryArticlesState createState() => _CategoryArticlesState();
+  State<CategoryArticles> createState() => _CategoryArticlesState();
 }
 
 class _CategoryArticlesState extends State<CategoryArticles> {
@@ -41,12 +41,10 @@ class _CategoryArticlesState extends State<CategoryArticles> {
 
   Future<List<dynamic>> fetchCategoryArticles(int page) async {
     try {
-      var response = await http.get(
-          Uri.parse("$WORDPRESS_URL/wp-json/wp/v2/posts?categories[]=" +
-              widget.id.toString() +
-              "&page=$page&per_page=10&_fields=id,date,title,content,custom,link"));
+      var response = await http.get(Uri.parse(
+          "$wordpressUrl/wp-json/wp/v2/posts?categories[]=${widget.id}&page=$page&per_page=10&_fields=id,date,title,content,custom,link"));
 
-      if (this.mounted) {
+      if (mounted) {
         if (response.statusCode == 200) {
           setState(() {
             categoryArticles.addAll(json
@@ -108,8 +106,9 @@ class _CategoryArticlesState extends State<CategoryArticles> {
         child: SingleChildScrollView(
             controller: _controller,
             scrollDirection: Axis.vertical,
-            child: Column(
-                children: <Widget>[categoryPosts(_futureCategoryArticles as Future<List<dynamic>>)])),
+            child: Column(children: <Widget>[
+              categoryPosts(_futureCategoryArticles as Future<List<dynamic>>)
+            ])),
       ),
     );
   }
@@ -119,12 +118,12 @@ class _CategoryArticlesState extends State<CategoryArticles> {
       future: categoryArticles,
       builder: (context, articleSnapshot) {
         if (articleSnapshot.hasData) {
-          if (articleSnapshot.data!.length == 0) return Container();
+          if (articleSnapshot.data!.isEmpty) return Container();
           return Column(
             children: <Widget>[
               Column(
                   children: articleSnapshot.data!.map((item) {
-                final heroId = item.id.toString() + "-categorypost";
+                final heroId = "${item.id}-categorypost";
                 return InkWell(
                   onTap: () {
                     Navigator.push(
@@ -141,7 +140,7 @@ class _CategoryArticlesState extends State<CategoryArticles> {
                   ? Container(
                       alignment: Alignment.center,
                       height: 30,
-                      )
+                    )
                   : Container()
             ],
           );
@@ -154,7 +153,6 @@ class _CategoryArticlesState extends State<CategoryArticles> {
         return Container(
           alignment: Alignment.center,
           height: 400,
-          
         );
       },
     );
